model/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
numpy>=1.20.0
opencv-python>=4.5.0
torch>=1.10.0
tensorflow>=2.8.0
facenet-pytorch>=2.5.2
scikit-learn>=1.0.0
scikit-image>=0.19.0
pymongo>=4.3.3
requests>=2.28.2
python-dotenv>=1.0.0